+++ include/H2DE/objects/H2DE_object.h
@@ -9,6 +9,13 @@ class H2DE_Object {
 private:
     bool hidden = false;
 
+    void update();
+    void updateCollision();
+    void snap(const H2DE_LevelRect& rect, const H2DE_LevelRect& otherRect, H2DE_Face face);
+
+    std::vector<H2DE_SurfaceBuffer> getSurfaceBuffers() const;
+    H2DE_Hitbox& getHitbox(const std::string& hitboxName);
+
 protected:
     H2DE_Engine* engine;
     H2DE_ObjectData od;
@@ -20,17 +27,11 @@ protected:
 
     static void destroySurfaces(std::unordered_map<std::string, H2DE_Surface*>& surfaces);
     static H2DE_Surface* getSurface(const std::unordered_map<std::string, H2DE_Surface*>& surfaces, const std::string& name);
-    
-    void update();
-    void updateCollision();
-    void snap(const H2DE_LevelPos& offset, const H2DE_LevelRect& rect, const H2DE_LevelRect& otherRect, H2DE_Face face);
 
     virtual void resetSurfaceBuffers() = 0;
     void clearSurfaceBuffers();
 
     static const std::vector<H2DE_Surface*> getSortedSurfaces(const std::unordered_map<std::string, H2DE_Surface*>& surfaces);
-    std::vector<H2DE_SurfaceBuffer> getSurfaceBuffers() const;
-    H2DE_Hitbox& getHitbox(const std::string& hitboxName);
 
     void addSurface(std::unordered_map<std::string, H2DE_Surface*>& surfaces, H2DE_Surface* surface, const std::string& name);
     void removeSurface(std::unordered_map<std::string, H2DE_Surface*>& surfaces, const std::string& name);
@@ -43,8 +44,9 @@ public:
 
     friend H2DE_LevelPos H2DE_GetObjectPos(const H2DE_Object* object);
     friend H2DE_LevelSize H2DE_GetObjectSize(const H2DE_Object* object);
-    friend float H2DE_GetObjectRotation(const H2DE_Object* object);
+    friend H2DE_LevelRect H2DE_GetObjectRect(const H2DE_Object* object);
     friend H2DE_LevelPos H2DE_GetObjectPivot(const H2DE_Object* object);
+    friend float H2DE_GetObjectRotation(const H2DE_Object* object);
     friend H2DE_Flip H2DE_GetObjectFlip(const H2DE_Object* object);
     friend std::unordered_map<std::string, H2DE_Hitbox> H2DE_GetObjectHitboxes(const H2DE_Object* object);
     friend int H2DE_GetObjectIndex(const H2DE_Object* object);
@@ -55,10 +57,12 @@ public:
     friend void H2DE_SetObjectPos(H2DE_Object* object, const H2DE_LevelPos& pos, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
     friend void H2DE_SetObjectSize(H2DE_Object* object, const H2DE_LevelSize& size);
     friend void H2DE_SetObjectSize(H2DE_Object* object, const H2DE_LevelSize& size, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
+    friend void H2DE_SetObjectRect(H2DE_Object* object, const H2DE_LevelRect& rect);
+    friend void H2DE_SetObjectRect(H2DE_Object* object, const H2DE_LevelRect& rect, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
     friend void H2DE_SetObjectIndex(H2DE_Object* object, int index);
+    friend void H2DE_SetObjectPivot(H2DE_Object* object, const H2DE_LevelPos& pivot);
     friend void H2DE_SetObjectRotation(H2DE_Object* object, float rotation);
     friend void H2DE_SetObjectRotation(H2DE_Object* object, float rotation, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
-    friend void H2DE_SetObjectPivot(H2DE_Object* object, const H2DE_LevelPos& pivot);
     friend void H2DE_SetObjectFlip(H2DE_Object* object, H2DE_Flip flip);
     friend void H2DE_SetObjectHitboxRect(H2DE_Object* object, const std::string& hitboxName, const H2DE_LevelRect& rect);
     friend void H2DE_SetObjectHitboxColor(H2DE_Object* object, const std::string& hitboxName, const H2DE_ColorRGB& color);